IdentifyFlash returns u32 now

+++ ginclude/gba/flash_internal.h
@@ -80,6 +80,6 @@ u16 ProgramFlashByte_MX(u16 sectorNum, u32 offset, u8 data);
 u16 ProgramFlashSector_MX(u16 sectorNum, void *src);
 
 // agb_flash_1m
-u16 IdentifyFlash(void);
+u32 IdentifyFlash(void);
 
 #endif // GUARD_GBA_FLASH_INTERNAL_H
